+++ src/main/java/org/rumbledb/items/StringItem.java
@@ -206,7 +206,7 @@ public class StringItem implements Item {
     @Override
     public boolean equals(Object otherItem) {
         if (otherItem instanceof Item) {
-            int c = ComparisonIterator.compareItems(
+            long c = ComparisonIterator.compareItems(
                 this,
                 (Item) otherItem,
                 ComparisonOperator.VC_EQ,
